+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.x.x] - Unreleased
+
+### Fixed
+
+- fix CopyAnchorButton: class -> className
+- extend HeaderAnchorDecorator from TooltipEntity
+- slugify custom header anchor
+- make all buttons in Tooltip small
+
+
 ## [0.0.1] - 2023-07-30
 
 ### Added